Move SignedTransaction into reth-primitives-traits
