--- crates/primitives/src/traits/transaction/mod.rs
@@ -1,3 +1,0 @@
-//! Transaction abstraction
-
-pub mod signed;
