+++ crates/primitives/src/transaction/signature.rs
@@ -16,29 +16,6 @@ const SECP256K1N_HALF: U256 = U256::from_be_bytes([
     0x5D, 0x57, 0x6E, 0x73, 0x57, 0xA4, 0x50, 0x1D, 0xDF, 0xE9, 0x2F, 0x46, 0x68, 0x1B, 0x20, 0xA0,
 ]);
 
-/// Decodes [`Signature`] w.r.t. chain ID.
-pub fn decode_with_eip155_chain_id(buf: &mut &[u8]) -> alloy_rlp::Result<(Signature, Option<u64>)> {
-    let v: Parity = Decodable::decode(buf)?;
-    let r: U256 = Decodable::decode(buf)?;
-    let s: U256 = Decodable::decode(buf)?;
-
-    #[cfg(not(feature = "optimism"))]
-    if matches!(v, Parity::Parity(_)) {
-        return Err(alloy_rlp::Error::Custom("invalid parity for legacy transaction"));
-    }
-
-    #[cfg(feature = "optimism")]
-    // pre bedrock system transactions were sent from the zero address as legacy
-    // transactions with an empty signature
-    //
-    // NOTE: this is very hacky and only relevant for op-mainnet pre bedrock
-    if matches!(v, Parity::Parity(false)) && r.is_zero() && s.is_zero() {
-        return Ok((Signature::new(r, s, Parity::Parity(false)), None))
-    }
-
-    Ok((Signature::new(r, s, v), v.chain_id()))
-}
-
 /// Recover signer from message hash, _without ensuring that the signature has a low `s`
 /// value_.
 ///
